requirements.txt
python-telegram-bot==21.10
python-dotenv
requests
